Separated annotation into directory
Split into separate files
